% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_J.R
\name{compute_J}
\alias{compute_J}
\title{compute_J}
\usage{
compute_J(K, lags)
}
\arguments{
\item{lag}{}
}
\value{

}
\description{
compute_J
}
\details{
Helper function to create the allocation matrix J, which can be used to transform a $VAR(p)$ into $VMA$ represntation.
}
