% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_Phi.R
\name{compute_Phi}
\alias{compute_Phi}
\title{compute_Phi}
\usage{
compute_Phi(lag, K, A_comp, const, n.ahead)
}
\arguments{
\item{n.ahead}{}
}
\value{

}
\description{
Helper function used to compute Phi coefficients as in Kilian, Luetkepohl
}
