% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_Theta.R
\name{compute_Theta}
\alias{compute_Theta}
\title{compute_Theta}
\usage{
compute_Theta(Phi, B_0)
}
\arguments{
\item{B_0}{}
}
\value{

}
\description{
Helper function used to compute Theta coefficients as in Kilian, Luetkepohl
}
