% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_theta_kj_sq.R
\name{compute_theta_kj_sq}
\alias{compute_theta_kj_sq}
\title{compute_theta_kj_sq}
\usage{
compute_theta_kj_sq(Theta, n.ahead)
}
\arguments{
\item{n.ahead}{}
}
\value{

}
\description{
compute_theta_kj_sq
}
\details{
Helper function used to compute orthogonalised IRFs as in Kilian, Luetkepohl. These represent the reduced form impulse response functions.
}
