% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deepvareg.R
\name{deepvareg}
\alias{deepvareg}
\title{Deep VAR}
\usage{
deepvareg(
  data,
  lags = 1,
  num_units = 50,
  num_layers = 2,
  p_drop_out = 0.5,
  horizon = 1,
  type = "var",
  verbose = 0,
  ...
)
}
\arguments{
\item{type}{}
}
\value{

}
\description{
Deep VAR
}
\author{
Patrick Altmeyer
}
