% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fevd.R
\name{fevd}
\alias{fevd}
\title{fevd}
\usage{
fevd(
  varresult,
  n.ahead = 10,
  plot = T,
  ylab = "Variance contribution",
  xlab = "Forecast horizon",
  title = NULL,
  theme = ggplot2::theme_bw()
)
}
\arguments{
\item{n.ahead}{}
}
\value{

}
\description{
Function used to compute forecast error variance decomposition as in Kilian, Luetkepohl
}
