% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/var_model.methods.R
\name{fitted.var_model}
\alias{fitted.var_model}
\title{Fitted values}
\usage{
\method{fitted}{var_model}(var_model, X = NULL)
}
\arguments{
\item{X}{}
}
\value{

}
\description{
Fitted values
}
\author{
Patrick Altmeyer
}
