% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/indentify_chol.R
\name{identify_chol}
\alias{identify_chol}
\title{identify_chol}
\usage{
identify_chol(Sigma_res)
}
\arguments{
\item{Sigma_res}{}
}
\value{

}
\description{
Computes coeffcient matrix capturing contemporaneous effects through short-run restriction using Cholesky decompostion.
}
