% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/irf.R
\name{irf}
\alias{irf}
\title{irf}
\usage{
irf(
  varresult,
  imp,
  structural = T,
  n_ahead = 10,
  size = 1,
  bootstrap = T,
  n_bootstrap = 1000,
  parametric = F,
  ci = 0.95,
  plot = T,
  cum = T
)
}
\arguments{
\item{size}{}
}
\value{

}
\description{
irf
}
\author{
Patrick Altmeyer
}
