% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare_deepvar_model.R
\name{prepare_deepvar_model}
\alias{prepare_deepvar_model}
\title{Deep VAR model setup}
\usage{
prepare_deepvar_model(deepvar_data, n_units = 50)
}
\arguments{
\item{n_units}{}
}
\value{

}
\description{
Deep VAR model setup
}
\author{
Patrick Altmeyer
}
