% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare_predictors.R
\name{prepare_predictors}
\alias{prepare_predictors}
\title{Title}
\usage{
prepare_predictors(data, lags)
}
\arguments{
\item{var_data}{}
}
\value{

}
\description{
Take a `var_data` object and prepares predictors for recursive forecasts.
}
