% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/red_ir.R
\name{red_ir}
\alias{red_ir}
\title{red_ir}
\usage{
red_ir(lag, K, A_comp, const, n.ahead)
}
\arguments{
\item{n.ahead}{}
}
\value{

}
\description{
Computes the reduced-form impulse responses. These are denoted as by Phi in Kilian, Luetkepohl.
}
