% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deepvar_model.methods.R
\name{train.deepvar_model}
\alias{train.deepvar_model}
\title{Title}
\usage{
\method{train}{deepvar_model}(deepvar_model, num_epochs = 50)
}
\arguments{
\item{num_epochs}{}
}
\value{

}
\description{
Title
}
