% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/VAR.R
\name{VAR}
\alias{VAR}
\title{VAR}
\usage{
VAR(data, lag, method = "ls", constant = T, ci = 0.95, verbose = T)
}
\arguments{
\item{data}{Data set}

\item{lag}{Number of lags}
}
\value{

}
\description{
VAR
}
\author{
Patrick Altmeyer
}
