% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/VAR_lag_select.R
\name{VAR_lag_select}
\alias{VAR_lag_select}
\title{VAR_lag_select}
\usage{
VAR_lag_select(data, max_lag = 10, ic_choice = "AIC")
}
\arguments{
\item{ic_choice}{}
}
\value{

}
\description{
VAR_lag_select
}
