% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/VAR_predict.R
\name{VAR_predict}
\alias{VAR_predict}
\title{Var_predict}
\usage{
VAR_predict(
  varresult,
  n.ahead = 10,
  ci = c(0.95, 0.75, 0.5),
  ci_colour = "darkblue",
  plot = F,
  dates = NULL,
  plot_vars = "all",
  linetype_fcast = "dashed",
  ylab = "Values",
  xlab = "Index",
  title = NULL,
  theme = ggplot2::theme_bw()
)
}
\arguments{
\item{theme}{}
}
\value{

}
\description{
Var_predict
}
