% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/VAR_stable.R
\name{VAR_stable}
\alias{VAR_stable}
\title{VAR_stable}
\usage{
VAR_stable(varresult, verbose = T)
}
\arguments{
\item{varresult}{Output from VAR estimation}

\item{verbose}{Logical: if TRUE, test result will be printed to console.}
}
\value{

}
\description{
VAR_stable
}
\author{
Patrick Altmeyer
}
