% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vareg.R
\name{vareg}
\alias{vareg}
\title{vareg}
\usage{
vareg(data, lags = 1, method = "ols", constant = TRUE, ci = 0.95)
}
\arguments{
\item{data}{Data set}

\item{lags}{Number of lags}
}
\value{

}
\description{
vareg
}
\author{
Patrick Altmeyer
}
