% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_theta_kj_sq.R
\name{compute_theta_kj_sq}
\alias{compute_theta_kj_sq}
\title{compute_theta_kj_sq}
\usage{
compute_theta_kj_sq(Theta, n.ahead = 10)
}
\arguments{
\item{n.ahead}{}
}
\value{

}
\description{
compute_theta_kj_sq
}
