% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deepvareg.R
\name{deepvareg}
\alias{deepvareg}
\title{Deep VAR}
\usage{
deepvareg(
  data,
  lags,
  response = NULL,
  n_ahead = 1,
  size_ensemble = 5,
  num_epochs = 50
)
}
\arguments{
\item{type}{}
}
\value{

}
\description{
Deep VAR
}
\author{
Patrick Altmeyer
}
