% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deepvareg.R
\name{deepvareg}
\alias{deepvareg}
\title{Deep VAR}
\usage{
deepvareg(
  train_ds,
  valid_ds,
  lags,
  response = NULL,
  n_ahead = 1,
  size_ensemble = 1,
  num_epochs = 100
)
}
\arguments{
\item{n_ahead}{}
}
\value{

}
\description{
Deep VAR
}
\author{
Patrick Altmeyer
}
